Dataset URL : https://bit.ly/3kXTdox
This is the third task of internship to predict different species of Iris flower based on it's features. There are only 6 feature columns in the dataset. It is instructed to use Decision tree algorithm which is a supervised learning algorithm. Moreover, I have tried to find optimal parameters to increase accuracy of model and decrease error in prediction.